+++ jwe/jwe.go
@@ -8,133 +8,12 @@ import (
 	"encoding/json"
 	"errors"
 	"log"
-	"net/url"
 
 	"github.com/lestrrat/go-jwx/buffer"
-	"github.com/lestrrat/go-jwx/internal/emap"
-	"github.com/lestrrat/go-jwx/jwa"
 )
 
 func debug(f string, args ...interface{}) {
 	log.Printf(f, args...)
-}
-
-func (h *Header) Set(key string, value interface{}) error {
-	switch key {
-	case "alg":
-		var v jwa.KeyEncryptionAlgorithm
-		s, ok := value.(string)
-		if ok {
-			v = jwa.KeyEncryptionAlgorithm(s)
-		} else {
-			v, ok = value.(jwa.KeyEncryptionAlgorithm)
-			if !ok {
-				return ErrInvalidHeaderValue
-			}
-		}
-		h.Algorithm = v
-	case "enc":
-		var v jwa.ContentEncryptionAlgorithm
-		s, ok := value.(string)
-		if ok {
-			v = jwa.ContentEncryptionAlgorithm(s)
-		} else {
-			v, ok = value.(jwa.ContentEncryptionAlgorithm)
-			if !ok {
-				return ErrInvalidHeaderValue
-			}
-		}
-		h.ContentEncryption = v
-	case "cty":
-		v, ok := value.(string)
-		if !ok {
-			return ErrInvalidHeaderValue
-		}
-		h.ContentType = v
-	case "kid":
-		v, ok := value.(string)
-		if !ok {
-			return ErrInvalidHeaderValue
-		}
-		h.KeyID = v
-	case "typ":
-		v, ok := value.(string)
-		if !ok {
-			return ErrInvalidHeaderValue
-		}
-		h.Type = v
-	case "x5t":
-		v, ok := value.(string)
-		if !ok {
-			return ErrInvalidHeaderValue
-		}
-		h.X509CertThumbprint = v
-	case "x5t#256":
-		v, ok := value.(string)
-		if !ok {
-			return ErrInvalidHeaderValue
-		}
-		h.X509CertThumbprintS256 = v
-	case "x5c":
-		v, ok := value.([]string)
-		if !ok {
-			return ErrInvalidHeaderValue
-		}
-		h.X509CertChain = v
-	case "crit":
-		v, ok := value.([]string)
-		if !ok {
-			return ErrInvalidHeaderValue
-		}
-		h.Critical = v
-	case "jku":
-		v, ok := value.(string)
-		if !ok {
-			return ErrInvalidHeaderValue
-		}
-		u, err := url.Parse(v)
-		if err != nil {
-			return ErrInvalidHeaderValue
-		}
-		h.JwkSetURL = u
-	case "x5u":
-		v, ok := value.(string)
-		if !ok {
-			return ErrInvalidHeaderValue
-		}
-		u, err := url.Parse(v)
-		if err != nil {
-			return ErrInvalidHeaderValue
-		}
-		h.X509Url = u
-	default:
-		h.PrivateParams[key] = value
-	}
-	return nil
-}
-
-func (h Header) MarshalJSON() ([]byte, error) {
-	return emap.MergeMarshal(h.EssentialHeader, h.PrivateParams)
-}
-
-func (h *Header) UnmarshalJSON(data []byte) error {
-	if h.EssentialHeader == nil {
-		h.EssentialHeader = &EssentialHeader{}
-	}
-	if h.PrivateParams == nil {
-		h.PrivateParams = map[string]interface{}{}
-	}
-
-	m := map[string]interface{}{}
-	if err := json.Unmarshal(data, &m); err != nil {
-		return err
-	}
-	for name, value := range m {
-		if err := h.Set(name, value); err != nil {
-			return err
-		}
-	}
-	return nil
 }
 
 func Parse(buf []byte) (*Message, error) {
@@ -237,38 +116,3 @@ func parseCompact(buf []byte) (*Message, error) {
 	}
 	return m, nil
 }
-
-/*
-func Encode(hdr, enckey, iv Base64Encoder, encrypt Encrypter) ([]byte, error) {
-
-// BASE64URL(UTF8(JWE Protected Header)) || '.' ||
-//   BASE64URL(JWE Encrypted Key) || '.' || BASE64URL(JWE Initialization
-//   Vector) || '.' || BASE64URL(JWE Ciphertext) || '.' || BASE64URL(JWE
-//   Authentication Tag).
-	parts := make([][]byte, 5)
-	hdrbuf, err := hdr.Base64Encode()
-	if err != nil {
-		return nil, err
-	}
-	parts[0] = hdrbuf
-
-	keybuf, err := enckey.Base64Encode()
-	if err != nil {
-		return nil, err
-	}
-	parts[1] = keybuf
-
-	ivbuf, err := iv.Base64Encode()
-	if err != nil {
-		return nil, err
-	}
-	parts[2] = ivbuf
-
-
-//	msg := NewMessage()
-//	msg.Crypter = append(msg.Crypter, NewRsaCrypt(contentalg, pubkey))
-
-
-	return nil, nil
-}
-*/
