+++ jwe/message.go
@@ -3,8 +3,11 @@ package jwe
 import (
 	"encoding/json"
 	"errors"
+	"net/url"
 
 	"github.com/lestrrat/go-jwx/buffer"
+	"github.com/lestrrat/go-jwx/internal/emap"
+	"github.com/lestrrat/go-jwx/jwa"
 )
 
 func NewRecipient() *Recipient {
@@ -18,6 +21,100 @@ func NewHeader() *Header {
 		EssentialHeader: &EssentialHeader{},
 		PrivateParams:   map[string]interface{}{},
 	}
+}
+
+func (h *Header) Set(key string, value interface{}) error {
+	switch key {
+	case "alg":
+		var v jwa.KeyEncryptionAlgorithm
+		s, ok := value.(string)
+		if ok {
+			v = jwa.KeyEncryptionAlgorithm(s)
+		} else {
+			v, ok = value.(jwa.KeyEncryptionAlgorithm)
+			if !ok {
+				return ErrInvalidHeaderValue
+			}
+		}
+		h.Algorithm = v
+	case "enc":
+		var v jwa.ContentEncryptionAlgorithm
+		s, ok := value.(string)
+		if ok {
+			v = jwa.ContentEncryptionAlgorithm(s)
+		} else {
+			v, ok = value.(jwa.ContentEncryptionAlgorithm)
+			if !ok {
+				return ErrInvalidHeaderValue
+			}
+		}
+		h.ContentEncryption = v
+	case "cty":
+		v, ok := value.(string)
+		if !ok {
+			return ErrInvalidHeaderValue
+		}
+		h.ContentType = v
+	case "kid":
+		v, ok := value.(string)
+		if !ok {
+			return ErrInvalidHeaderValue
+		}
+		h.KeyID = v
+	case "typ":
+		v, ok := value.(string)
+		if !ok {
+			return ErrInvalidHeaderValue
+		}
+		h.Type = v
+	case "x5t":
+		v, ok := value.(string)
+		if !ok {
+			return ErrInvalidHeaderValue
+		}
+		h.X509CertThumbprint = v
+	case "x5t#256":
+		v, ok := value.(string)
+		if !ok {
+			return ErrInvalidHeaderValue
+		}
+		h.X509CertThumbprintS256 = v
+	case "x5c":
+		v, ok := value.([]string)
+		if !ok {
+			return ErrInvalidHeaderValue
+		}
+		h.X509CertChain = v
+	case "crit":
+		v, ok := value.([]string)
+		if !ok {
+			return ErrInvalidHeaderValue
+		}
+		h.Critical = v
+	case "jku":
+		v, ok := value.(string)
+		if !ok {
+			return ErrInvalidHeaderValue
+		}
+		u, err := url.Parse(v)
+		if err != nil {
+			return ErrInvalidHeaderValue
+		}
+		h.JwkSetURL = u
+	case "x5u":
+		v, ok := value.(string)
+		if !ok {
+			return ErrInvalidHeaderValue
+		}
+		u, err := url.Parse(v)
+		if err != nil {
+			return ErrInvalidHeaderValue
+		}
+		h.X509Url = u
+	default:
+		h.PrivateParams[key] = value
+	}
+	return nil
 }
 
 func (h1 *Header) Copy(h2 *Header) error {
@@ -46,6 +143,30 @@ func (h1 *Header) Copy(h2 *Header) error {
 		h1.PrivateParams[k] = v
 	}
 
+	return nil
+}
+
+func (h Header) MarshalJSON() ([]byte, error) {
+	return emap.MergeMarshal(h.EssentialHeader, h.PrivateParams)
+}
+
+func (h *Header) UnmarshalJSON(data []byte) error {
+	if h.EssentialHeader == nil {
+		h.EssentialHeader = &EssentialHeader{}
+	}
+	if h.PrivateParams == nil {
+		h.PrivateParams = map[string]interface{}{}
+	}
+
+	m := map[string]interface{}{}
+	if err := json.Unmarshal(data, &m); err != nil {
+		return err
+	}
+	for name, value := range m {
+		if err := h.Set(name, value); err != nil {
+			return err
+		}
+	}
 	return nil
 }
